requirements.txt
bleach==6.0.0
click==8.1.3
email-validator==1.3.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
python-magic==0.4.27
SQLAlchemy==2.0.4
Werkzeug==2.2.3
WTForms==3.0.1
